backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2

# Optional: For RAG/Vector Search (uncomment if needed)
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4

# Optional: For local HuggingFace transformers (uncomment if needed)
# transformers==4.35.0
# torch==2.1.0
# sentencepiece==0.1.99
# accelerate==0.24.0